[native] Move sessionInvalidationAlert to redux-setup.js
